arreglando bugs primera sucursal
