+++ src/views/configuracion.vue
@@ -91,6 +91,7 @@
 <script>
 import axios from 'axios'
 import cookies from 'vue-cookies';
+import { alertQuestion } from '../components/alert.vue';
 import { mapState, mapActions } from 'vuex';
 // exportar componente
 export default {
@@ -121,11 +122,13 @@ export default {
         setUsuario(usuario) {
             this.actionUsuario(usuario);
         },
-        cerrarSesion() {
-            if (confirm('Desea cerrar sesión?')) {
+        async cerrarSesion() {
+            if (await alertQuestion('Desea cerrar sesión?', null, 'Aceptar', null, null, true)) {
                 this.$cookies.remove('auth');
                 localStorage.removeItem('auth')
+                this.$router.push('/')
             }
+
         },
         // método para redireccionar a página anterior
         back() {
